requirements.txt
streamlit>=1.28
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.28.0
scipy>=1.10.0
plotly>=5.0.0
pyarrow>=12.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
pyxlsb>=1.0.10
